add test for fetch_track_metadata_for_items
